+++ source/mundo/Hijo.java
@@ -5,8 +5,8 @@ import java.util.Date;
 
 public class Hijo extends Persona implements Serializable{
 		
-	public Hijo(String nombre, String apellidos, int telefono,
-			int identificacion, String sexo, String direccion, String ciudad, String departamento,
+	public Hijo(String nombre, String apellidos, double telefono,
+			double identificacion, String sexo, String direccion, String ciudad, String departamento,
 			String tipoDocumentoP, Date fechaNacimientoP) {
 		
 		super(identificacion, tipoDocumentoP, nombre, apellidos, sexo, fechaNacimientoP,
